Fixed bug, greatly improved performance
Fixed a bug causing solutions to be left out if they included a piece that's split twice, but if the top split was placed with the middle split it would overlap gray minos (in other words, when it was seeing if it could place the top split of a piece, it tested the entire top split of a piece against the gray minos instead of just the row it was trying to place). For example, this solution would not have been found: https://fumen.zui.jp/?v115@zgE8h0R8g0S8g0D8JeAgH

An adjusted version of the dynamic programming system in the solution checker (when pruning impossible solutions) has been re-implemented. The original version had to be disabled because it wasn't compatible with piece restrictions, but now it considers the current piece options and the current hold options, making it functional and most likely just as efficient as the original. This drastically decreases the runtime for inputs with many options for pieces (eg: -p *p1,*p1,*p1). For example, `sfaster -p [OI]p1[OI]p1[OI]p1[OI]p1[OI]p1[OI]p1[OI]p1[OI]p1[OI]p1[OI]p1` took my device 6:42s to run, now it takes 0.147s to run.