Dramatically increasing speed!
There was a bug with the DP map (meant to prune redundant paths when checking solutions) that was keeping it from catching a lot of redundant paths, aka it was checking a lot of piece sequences it should've already known wouldn't work. From the handful of speed tests I ran comparing pre- and post-fix, total runtime can be 2x smaller or more! Theoretically, the speed increase should be more significant the more pieces there are.

(To be clear, the old version still found every solution- the bug just slowed it down.)